+++ R/group.R
@@ -97,6 +97,7 @@ group_exec_int <- function(
 
     status <- run_group_int(
       handle[['name']],
+      preproc[['fun.name']],
       alp[['dat']],
       alp[['dat_cols']],
       alp[['ids']],
@@ -279,9 +280,19 @@ r2c_groups_template <- function() {
 #' r2c_mean <- r2cq(mean(x))
 #' with(mtcars, group_exec(r2c_mean, hp, groups=cyl))
 #'
+#' ## Slope of `y ~ x`
 #' r2c_slope <- r2cq(
 #'   sum((x - mean(x)) * (y - mean(y))) / sum((x - mean(x)) ^ 2)
 #' )
+#' ## Equivalently as follows, but `r2c` recognizes and optimizes
+#' ## re-used calls (see `?reuse_calls`) so the intermediate assignments
+#' ## are not necessary.
+#' \dontrun{
+#' r2c_slope <- r2cq({
+#'   mux <- mean(x)
+#'   x_mux <- x - mux
+#'   sum(x_mux * (y - mean(y))) / sum(x_mux^2)
+#' })}
 #' with(mtcars, group_exec(r2c_slope, list(hp, qsec), groups=cyl))
 #'
 #' ## Parameters are generated in the order they are encountered
@@ -355,11 +366,12 @@ group_exec <- function(fun, data, groups, MoreArgs=list()) {
 }
 
 run_group_int <- function(
-  handle, dat, dat_cols, ids, extern, group.sizes, group.res.sizes
+  handle, fun.name, dat, dat_cols, ids, extern, group.sizes, group.res.sizes
 ) {
   .Call(
     R2C_run_group,
     handle,
+    fun.name,
     dat,
     dat_cols,
     ids,